+++ spec/models/rating_spec.rb
@@ -7,6 +7,6 @@ describe Yt::Rating do
   describe '#update' do
     before { expect(rating).to receive(:do_update).and_yield }
 
-    it { expect{rating.update :like}.to change{rating.rating} }
+    it { expect{rating.set :like}.to change{rating.rating} }
   end
 end