+++ HISTORY.md
@@ -2,6 +2,7 @@ v0.8 - 2014/07/18
 -----------------
 
 * [breaking change] channel.subscribe returns nil (not raise an error) when trying to subscribe to your own channel
+* Add all the status fields to Video (upload status, failure reason, rejection reason, scheduled time, license, embeddable, public stats viewable)
 
 v0.7 - 2014/06/18
 -----------------